# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signin": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User signed in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResponseDto"
                }
              }
            }
          }
        },
        "summary": "Sign in to an existing user account",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/signup": {
      "post": {
        "description": "Endpoint to create a new user account",
        "operationId": "AuthController_createNewUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User account created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "msg": "Signup successful"
                  }
                }
              }
            }
          }
        },
        "summary": "Create a new user account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Endpoint to refresh access token, send refresh token in cookie",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Access token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJ1dWlkLTEyMzQiLCJpYXQiOjE2ODgwODc2MDAsImV4cCI6MTY4ODA5MTIwMH0.XYZ"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "refreshToken": []
          },
          {
            "refreshToken": []
          }
        ],
        "summary": "Refresh access token using refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/friends": {
      "get": {
        "operationId": "FriendsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2025-11-14T12:00:00Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FriendshipResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List friends for current user",
        "tags": [
          "Friends"
        ]
      }
    },
    "/friends/are-friends/{userId}": {
      "get": {
        "operationId": "FriendsController_areFriends",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Check if current user is friends with another user with userId",
        "tags": [
          "Friends"
        ]
      }
    },
    "/friends/{otherUserId}": {
      "delete": {
        "operationId": "FriendsController_remove",
        "parameters": [
          {
            "name": "otherUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Remove a friend (delete friendship)",
        "tags": [
          "Friends"
        ]
      }
    },
    "/friend-requests": {
      "post": {
        "operationId": "FriendRequestsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFriendRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendRequestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Create friend request",
        "tags": [
          "Friend Requests"
        ]
      }
    },
    "/friend-requests/to-me": {
      "get": {
        "operationId": "FriendRequestsController_list",
        "parameters": [
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "example": "incoming",
              "type": "string",
              "enum": [
                "incoming",
                "outgoing"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "pending",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FriendRequestResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List friend requests sent to me",
        "tags": [
          "Friend Requests"
        ]
      }
    },
    "/friend-requests/from-me": {
      "get": {
        "operationId": "FriendRequestsController_listFromMe",
        "parameters": [
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "example": "incoming",
              "type": "string",
              "enum": [
                "incoming",
                "outgoing"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "pending",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FriendRequestResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List friend requests sent from me",
        "tags": [
          "Friend Requests"
        ]
      }
    },
    "/friend-requests/{id}/accept": {
      "post": {
        "operationId": "FriendRequestsController_accept",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Accept a friend request",
        "tags": [
          "Friend Requests"
        ]
      }
    },
    "/friend-requests/{id}/reject": {
      "post": {
        "operationId": "FriendRequestsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Reject a friend request",
        "tags": [
          "Friend Requests"
        ]
      }
    },
    "/friend-requests/{id}": {
      "delete": {
        "operationId": "FriendRequestsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Cancel a friend request",
        "tags": [
          "Friend Requests"
        ]
      }
    },
...
  },
  "info": {
    "title": "BlauChat API Documentation",
    "description": "The description of my application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3600/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      },
      "accessToken": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "SignInDto": {
        "type": "object",
        "properties": {}
      },
      "SignInResponseDto": {
        "type": "object",
        "properties": {}
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "title": "First Name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "title": "Last Name",
            "example": "Doe"
          },
          "username": {
            "type": "string",
            "title": "Username",
            "example": "johndoe"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "example": "johndoe@example.com"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "example": "strongPassword123"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "username",
          "email",
          "password"
        ]
      },
...
      "FriendshipResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-1"
          },
          "user_id": {
            "type": "string",
            "example": "uuid-user"
          },
          "friend_id": {
            "type": "string",
            "example": "uuid-friend"
          },
          "created_at": {
            "type": "string",
            "example": "2025-11-14T12:00:00Z"
          }
        },
        "required": [
          "id",
          "user_id",
          "friend_id",
          "created_at"
        ]
      },
      "CreateFriendRequestDto": {
        "type": "object",
        "properties": {
          "toUserId": {
            "type": "string",
            "description": "User id to send request to",
            "example": "uuid-1234"
          },
          "message": {
            "type": "string",
            "description": "Optional message",
            "example": "Hi, let's connect"
          }
        },
        "required": [
          "toUserId"
        ]
      },
      "FriendRequestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-1"
          },
          "fromUserId": {
            "type": "string",
            "example": "uuid-from",
            "description": "user id who sent the request"
          },
          "toUserId": {
            "type": "string",
            "example": "uuid-to",
            "description": "user id who receives the request"
          },
          "status": {
            "type": "string",
            "example": "pending"
          },
          "message": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "example": "2025-11-14T12:00:00Z"
          }
        },
        "required": [
          "id",
          "fromUserId",
          "toUserId",
          "status",
          "createdAt"
        ]
      },
...
    }
  }
}